% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_ajsurvplot.R
\name{gg_ajsurvplot}
\alias{gg_ajsurvplot}
\title{Plot the Aalen-Johansen curve with ggplot for competing risks}
\usage{
gg_ajsurvplot(
  formula,
  data,
  weights,
  subset,
  na.action,
  etype,
  count,
  id,
  timefix,
  backend = c("ggsurvfit", "survminer"),
  ...
)
}
\arguments{
\item{formula}{A standard model formula, with survival on the left and covariates on the right}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model}

\item{weights, subset, na.action, }{count,id,timefix parameter passed to finegray model. See \code{\link[survival:finegray]{finegray}},}

\item{backend}{[\code{ggsurvfit}] either 'survminer' or 'ggsurvfit'}

\item{...}{optional parameters passed to \code{ggsurvplot} or ].}

\item{event}{event of interest, passed to finegray as etype.}
}
\description{
Extend \link[survminer:ggsurvplot]{ggsurvplot} functionality to plot competing risk curve
}
