% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_ajsurvplot.R
\name{ggsurvfit2}
\alias{ggsurvfit2}
\title{ggsurvfit with strata saved}
\usage{
ggsurvfit2(
  x,
  type = "survival",
  linetype_aes = FALSE,
  theme = theme_ggsurvfit_default(),
  ...
)
}
\arguments{
\item{x, type, linetype_aes, theme, ...}{parameters passed directly \link[ggsurvfit:ggsurvfit]{ggsurvfit}.}
}
\description{
Same as \link[ggsurvfit:ggsurvfit]{ggsurvfit} but with strata saved, so you can facet by covariables.
}
