% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy_survfit2}
\alias{tidy_survfit2}
\title{Tidy survfit with strata save}
\usage{
tidy_survfit2(x, times = NULL, type = "survival")
}
\arguments{
\item{x, times, type}{same as \link[ggsurvfit:tidy_survfit]{tidy_survfit}}
}
\description{
Same as \link[ggsurvfit:tidy_survfit]{tidy_survfit} but with strata saved.
}
